move tool_context_config to lab_setting.toml
