+++ src/lab/mcp/__init__.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from ._typing import ConversationState, ScreenShotResult, ToolCallLike, ToolContextConfig, ToolMessage, ToolTraceItem
+from ._typing import ConversationState, ScreenShotResult, ToolCallLike, ToolMessage, ToolTraceItem
 from .context_policy import build_tool_context
 from .fastmcp_router import FastMcpRouter
 from .tool_registry import DEFAULT_RETRY_HINT, TOOL_RETRY_HINTS, ToolRegistry
@@ -15,6 +15,5 @@ __all__ = [
     "DEFAULT_RETRY_HINT",
     "TOOL_RETRY_HINTS",
     "ConversationState",
-    "ToolContextConfig",
     "build_tool_context",
 ]
